drop useless test
